Generator rework, added controller structure
The GenerateFractals module has been reworked into Generator; function definitions changed to a form that allows easy use in other modules.

The primary control flow has been added to the stepHandler -- this does not yet work.

A few minor helper definitions are added to the Model

the to-be-implemented getColours function has been added to View