Expanded the GeneratorData type, added a Fractal enumeration type, changed layout
Generator Data now keeps track of the complex coordinates z and the complex parameter c for each fractal (type). Degree represents the degree of the polynomial z^{degree}. In our world state, GenData is initialised with (0,0) for both z and c, and a degree of 2. I added some questions that I couldn't answer as well in comment format.