Update ZFunction, adjusted functions using ZFunction
Rewritten ZFunction to accept more parameters and be able to differentiate between different fractals.   Renamed nextZDefault to nextFractalPoint and rewritten the code so that it can be extended to different fractal computations based on the input fractal type. Still using only the Mandelbrot polynomial. Changed getIterations, fixedC, fixedZ and getSequences to work with the new GeneratorData type and ZFunction type signature.

Added some comments with questions about our design.